requirements.txt
Flask>=3.0,<4.0
PyExecJS>=1.5,<2.0
playwright>=1.58,<2.0
